+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Add `DeviceLoaderBuilder::build_with_existing_device`.
 - Add `#[must_use]` attribute to builder methods
 
 ### Changed